# changed default from ball_tree to auto. Should be faster now for larger datasets